backend/requirements.txt
# Production dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0

# Async database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP client
httpx==0.27.2

# OpenAI
openai==1.54.0

# AWS S3 / Cloudflare R2
boto3==1.35.0
aioboto3==13.2.0

# Task queue
celery[redis]==5.4.0
redis==5.2.0
flower==2.0.1

# Stripe
stripe==11.2.0

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0         # Retry logic for external API calls
orjson==3.10.11         # Fast JSON

# Monitoring
sentry-sdk[fastapi]==2.18.0

# Dev only
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
factory-boy==3.3.1
